+++ src/svgr.d.ts
@@ -2,3 +2,8 @@ declare module "*.svg" {
   const value: React.ComponentType<React.SVGAttributes<SVGElement>>;
   export default value;
 }
+
+declare module "*.png" {
+  const value: string;
+  export default value;
+}