+++ src/types/dtos/jwt-token-payload.ts
@@ -3,7 +3,6 @@ import { Roles } from "../roles";
 export interface JwtTokenPayload {
   id: string;
   username: string;
-  displayname: string;
   role: Roles;
   isAdmin: boolean;
   childrenIds: string[];