+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [v0.40.4] - 2020-09-05
+- Split "display name" into "first name" and "last name"
+
+## [v0.40.4] - 2020-09-05
+- Allow filtering by class
+- Add 404 Page
 
 - Added username to invitation link
 - Persist number of rows per page